move module declarations to separate file
