+++ mui.d.ts
@@ -1,0 +1,43 @@
+import '@mui/material/InputBase'
+declare module '@mui/material/InputBase' {
+  // eslint-disable-next-line no-unused-vars
+  interface InputBasePropsColorOverrides {
+    'mvc-green': true
+    'mvc-yellow': true
+    'mvc-white': true
+    'mvc-gray': true
+  }
+}
+
+import '@mui/material/TextField'
+declare module '@mui/material/TextField' {
+  // eslint-disable-next-line no-unused-vars
+  interface TextFieldPropsColorOverrides {
+    'mvc-green': true
+    'mvc-yellow': true
+    'mvc-white': true
+    'mvc-gray': true
+  }
+}
+
+import '@mui/material/CircularProgress'
+declare module '@mui/material/CircularProgress' {
+  // eslint-disable-next-line no-unused-vars
+  interface CircularProgressPropsColorOverrides {
+    'mvc-green': true
+    'mvc-yellow': true
+    'mvc-white': true
+    'mvc-gray': true
+  }
+}
+
+import '@mui/material/Button'
+declare module '@mui/material/Button' {
+  // eslint-disable-next-line no-unused-vars
+  interface ButtonPropsColorOverrides {
+    'mvc-green': true
+    'mvc-yellow': true
+    'mvc-white': true
+    'mvc-gray': true
+  }
+}
